+++ CHANGELOG.md
@@ -10,6 +10,20 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 <!-- TODO: change `@added unknown` tag due to the actual version -->
 <!--       during new scriptlets or redirects releasing -->
 
+## [Unreleased]
+
+### Added
+
+- New `mode` parameter to all `json-prune` and `trusted-json-set` related scriptlets,
+  which allows to specify the way of pruning JSON objects [#522].
+- Line-delimited JSON processing in `trusted-json-set` related scriptlets [#522].
+- `$remove$` value to all `trusted-json-set` related scriptlets, which allows removing specific properties [#522].
+- Support for `JSONPath` in `json-prune`, `json-prune-fetch-response`, `json-prune-xhr-response`,
+  `trusted-json-set`, `trusted-json-set-fetch-response` and `trusted-json-set-xhr-response` scriptlets [#522].
+
+[Unreleased]: https://github.com/AdguardTeam/Scriptlets/compare/v2.3.1...HEAD
+[#522]: https://github.com/AdguardTeam/Scriptlets/issues/522
+
 ## [v2.3.1] - 2026-03-24
 
 ### Changed